+++ documentation/ChangeLog.md
@@ -1,6 +1,7 @@
 ## release ChangeLog
 
 ### (NextVersion)
+ * Fixed: if build tool reports many error messages, the Eclipse project build will take too long to finish. (#26)
  * Added UserGuide note about using Homebrew GDB in OS X.
  * Fixed: "Run As"/"Debug As" launch shortcut incorrectly matching pre-existing launch configurations.
  * Fixed: project not being refreshed when Build Target build invoked directly from Project Explorer.